requirements.txt
CherryPy==3.2.4
Magic-file-extensions==0.2
PyPerf==0.1.0
Pygments==1.6
Scrapy==0.20.0
Twisted==13.2.0
binwalk==1.2.2-1
bunny1==1.1
cheat==1.0
coverage==3.7
cssselect==0.9.1
dom==0.2.2
elasticsearch==0.4.3
fu==0.1.1
lxml==3.2.4
mercurial==2.9
nose==1.3.0
numpy==1.8.0
pyOpenSSL==0.13.1
pyobjc==2.5.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-CalendarStore==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServerNotification==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
python-ldap==2.4.14
queuelib==1.1.1
requests==1.2.3
simplejson==3.3.0
six==1.4.1
termcolor==1.1.0
urllib3==1.7.1
vboxapi==1.0
virtualenv==1.11.2
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.0.5
